Split dispatch.rs into the module tier, the return sync and the heap floor parts
